# uint256: optimize ExtendSign