close #102

## Test
```
go test ./...
```
returns
```
ok  	github.com/holiman/uint256	0.976s
```
## Benchmark
code:
```
func BenchmarkExtendSign(bench *testing.B) {
	a := new(Int).SetBytes(hex2Bytes("f123456789abcdeffedcba9876543210f2f3f4f5f6f7f8f9fff3f4f5f6f7f8f9"))

	result := new(Int)
	n := NewInt(0)
	bench.ResetTimer()
	for i := 0; i < bench.N; i++ {
		result.ExtendSign(a, n)
	}
}
```
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
              │     old     │                 new                 │
              │   sec/op    │   sec/op     vs base                │
ExtendSign-16   7.875n ± 1%   2.859n ± 1%  -63.70% (p=0.000 n=10)

              │    old     │              new               │
              │    B/op    │    B/op     vs base            │
ExtendSign-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

              │    old     │              new               │
              │ allocs/op  │ allocs/op   vs base            │
ExtendSign-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```