@holiman 2024-05-06
I added a benchmark:
```
[user@work uint256]$ benchstat bench.extend.1 bench.extend.2
name          old time/op  new time/op  delta
ExtendSign-8  13.0ns ±22%  12.0ns ± 6%  -7.35%  (p=0.030 n=8+10)
```

I didn't use the same one you did: yours was skewed to only ever bench with `n=0`. So my speedup is a bit less, I'd be curious to see if you get roughly the same improvements as me with this benchmark, or if you still get something much larger?