@AaronChen0 2024-05-06
Using your benchmark, I still get much better performance.
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
              │     old     │                 new                 │
              │   sec/op    │   sec/op     vs base                │
ExtendSign-16   8.175n ± 2%   3.557n ± 1%  -56.49% (p=0.000 n=10)

              │    old     │              new               │
              │    B/op    │    B/op     vs base            │
ExtendSign-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

              │    old     │              new               │
              │ allocs/op  │ allocs/op   vs base            │
ExtendSign-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```