@holiman 2024-05-06
> What system are you testing on?

It's a lenovo x1 carbon gen 10, `12th Gen Intel(R) Core(TM) i7-1270P1`, but it's a bit special because I'm running on qubes os, so it's on a virtual machine.